Chaning ViewCount to long.

+++ SpeedRunAppImport.Interfaces/Services/IScrapeService.cs
@@ -6,6 +6,6 @@ namespace SpeedRunAppImport.Interfaces.Services
     public interface IScrapeService
     {
         IEnumerable<string> GetLatestSpeedRunIDs();
-        int? GetYouTubeViewCount(string videoLinkUrl);
+        long? GetYouTubeViewCount(string videoLinkUrl);
     }
 }
